feat: add download resource file CLI command
