+++ udata_hydra/utils/__init__.py
@@ -1,7 +1,13 @@
 from .auth import token_auth_middleware
 from .csv import detect_tabular_from_headers
 from .errors import IOException, ParseException, handle_parse_exception
-from .file import compute_checksum_from_file, download_resource, extract_gzip, remove_remainders
+from .file import (
+    compute_checksum_from_file,
+    download_file,
+    download_resource,
+    extract_gzip,
+    remove_remainders,
+)
 from .geojson import detect_geojson_from_headers_or_catalog
 from .http import UdataPayload, get_request_params, send
 from .queue import enqueue
@@ -14,6 +20,7 @@ __all__ = [
     "ParseException",
     "handle_parse_exception",
     "compute_checksum_from_file",
+    "download_file",
     "download_resource",
     "extract_gzip",
     "remove_remainders",
